Refactor some gitHub vs git functions
