+++ src/tree/ActionTreeItem.ts
@@ -9,7 +9,7 @@ import { AzExtTreeItem, AzureParentTreeItem, IActionContext, TreeItemIconPath } 
 import { createOctokitClient } from '../commands/github/createOctokitClient';
 import { ActionsGetWorkflowRunResponseData, ActionsListJobsForWorkflowRunResponseData } from '../gitHubTypings';
 import { ensureStatus, getActionIconPath } from '../utils/actionUtils';
-import { getRepoFullname } from '../utils/gitHubUtils';
+import { getRepoFullname } from '../utils/gitUtils';
 import { ActionsTreeItem } from "./ActionsTreeItem";
 import { IAzureResourceTreeItem } from './IAzureResourceTreeItem';
 import { JobTreeItem } from './JobTreeItem';
